Fixed a bug where clearing a placeholder would remove the definition entirely, instead of just setting the draft definition to null.
git-svn-id: https://smartest-php5.repos.vscmedia.com@684 3b18a885-9e44-0410-b831-aac8df1b01cb

+++ System/Applications/Pages/Pages.class.php
@@ -3076,7 +3076,7 @@ class Pages extends SmartestSystemApplication{
 	            
 	            $definition = new SmartestPlaceholderDefinition;
 	            
-	            if($definition->loadForUpdate($placeholder->getName(), $page, $item_id)){
+	            if(is_numeric($item_id) && $definition->loadForUpdate($placeholder->getName(), $page, $item_id)){
 	                
 	                // update placeholder
 	                $definition->delete();
@@ -3085,9 +3085,9 @@ class Pages extends SmartestSystemApplication{
 	            }else if($definition->loadForUpdate($placeholder->getName(), $page)){
 	                
 	                // update placeholder
-	                $definition->setDraftAssetId('');
+	                $definition->setDraftAssetId(null);
 	                $definition->save();
-	                $this->addUserMessageToNextRequest('The placeholder definition was removed.', SmartestUserMessage::SUCCESS);
+	                $this->addUserMessageToNextRequest('The draft placeholder definition was removed. Next time the page is published, the placeholder will be emptied.', SmartestUserMessage::SUCCESS);
 	                
 	            }else{
